admin backend rough complete

+++ blog.sql
@@ -53,7 +53,7 @@ create table if not exists `Resource`(
     `resource_id` int auto_increment,
     `user_type` tinyint(1),
     `file_name` varchar(128),
-    `file_type` tinyint(1),
+    `file_type` int,
     `url` varchar(255),
     `time` datetime not null default now(),
     primary key (`resource_id`)
